mq_timedsend01: Fix different signedness error on 32bit
mq_timedsend01.c:223:21: warning: comparison of integer expressions of
different signedness: ‘unsigned int’ and ‘long int’ [-Wsign-compare]

Link: https://lore.kernel.org/ltp/20241114144029.349559-2-pvorel@suse.cz/
Acked-by: Jan Stancek <jstancek@redhat.com>

+++ testcases/kernel/syscalls/mq_timedsend/mq_timedsend01.c
@@ -235,7 +235,7 @@ static void do_test(unsigned int i)
 		}
 	}
 
-	if (tc->len != TST_RET) {
+	if ((long)tc->len != TST_RET) {
 		tst_res(TFAIL, "mq_timedreceive() wrong length %ld, expected %u",
 			TST_RET, tc->len);
 		return;
